requirements.txt
# Chatterbox TTS Service - Dependencies
# Pin versions for reproducibility

# =============================================================================
# Core Framework
# =============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
sse-starlette==1.8.2
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0

# =============================================================================
# ML/Deep Learning
# =============================================================================
torch==2.1.2
torchaudio==2.1.2
safetensors==0.4.1
transformers==4.36.2
accelerate==0.25.0

# Chatterbox TTS
chatterbox-tts>=0.1.0

# =============================================================================
# Optimization - ONNX/TensorRT
# =============================================================================
onnx==1.15.0
onnxruntime-gpu==1.16.3
# tensorrt - install via NVIDIA container or pip install tensorrt==10.0.0

# vLLM - CRITICAL: Must be exactly 0.9.2 for Chatterbox compatibility
vllm==0.9.2

# =============================================================================
# Audio Processing
# =============================================================================
numpy==1.26.2
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1
resampy==0.4.2

# =============================================================================
# Caching & Storage
# =============================================================================
redis==5.0.1
aioredis==2.0.1
huggingface-hub==0.20.1

# =============================================================================
# Monitoring & Logging
# =============================================================================
prometheus-client==0.19.0
python-json-logger==2.0.7

# =============================================================================
# Testing & Development
# =============================================================================
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.25.2
locust==2.20.0

# =============================================================================
# Utilities
# =============================================================================
click==8.1.7
tqdm==4.66.1
python-dotenv==1.0.0
